requirements-core.txt
torch>=2.0.0
numpy
scipy
scikit-learn
pandas
matplotlib
PyYAML
zarr>=2.12
tqdm
